refactor: changed data type INT -> BIGINT

+++ backend/services/product-service/src/main/resources/db/migration/V11__create_product_item_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE product_item (
     id BIGSERIAL PRIMARY KEY,
-    size INT,
+    size BIGINT,
     price DECIMAL(5,2) NOT NULL,
     dough_type_id BIGINT,
     product_id BIGINT NOT NULL
